requirements.txt
Django~=3.2.8
django-filter~=21.1
django-widget-tweaks~=1.4.8
Pillow~=8.3.2
